Restructure core classes
I wanted methods to manipulate a set of commits. The Person class
consisted of a name, a set of commits, and methods to calculate
statistics on that set of commits. I created a new class, CommitSet, to
house those methods and moved the relevant methods from Person to
CommitSet. That obviated the need for the Person class and I removed it.

CommitStats#get_person and CommitStats#people_by_most_commits were
methods on a set of commits so, I moved them to CommitSet#select_author
and CommitSet#top_authors. After that, CommitStat only had one method in
it that gathered commits from multiple repositories, so I gave it a more
appropriate name and made it a module.

I used to remove commits made by people outside @causes.com. I removed
that code because it's specific to our organization and because I want
to see contributions from members of the open-source community.

Change-Id: If688a43180797279279269945511ebe5df8ecde4
Reviewed-on: https://gerrit.causes.com/22012
Reviewed-by: Jimmy Kittiyachavalit <jimmy@causes.com>
Tested-by: Jimmy Kittiyachavalit <jimmy@causes.com>