Replace segfaulting Rugged gem with the magic of regexen
This is not quite as performant as Rugged, but it works and doesn't
segfault. The regex magic is a fiddly, but that is the nature of
parsing; we're relying on "plumbing" formats, so we can be confident
that Git's output will be stable and our parsing will continue to work.
In this particular case we don't have to worry about nesting or
recursion -- the language we are parsing is effectively a "regular
language" -- so a regex is a good, cheap fit.

We removed the `rescue` clause because we wanted things to fail loudly
during development. Running this code over all of our repos for 300
days' worth of data shows that things work, edge cases included, so we
can probably do without the `rescue` forever.

Change-Id: I890b528de3b68335fb03f803f62814611f652d74
Reviewed-on: https://gerrit.causes.com/24876
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>