Memoize Twig::Repo#commits results
Retrieving the commits for a repo is a potentially costly operation, so
memoize the results, keyed on the `options`. We don't currently hit this
method multiple times in a specific run, but we will likely want to do
so for the web app in the near future; we might want some notion of
expiry as well, but we'll add that when and if we need it.

Change-Id: I03a6ab1a12cd455fef4e9f00b00a502bbf3ac5ae
Reviewed-on: https://gerrit.causes.com/24881
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>