Use Pathname#basename to extract repo name
This isolates us from platform-specific details like whether "/" is
actually a path separator or not; I believe that Pathname will render
path separators as slashes even on Windows, but there is no guarantee
that that will always be the case so we might as well be future-proof.

As a bonus, the code is shorter and arguably more intention-revealing.

Change-Id: I6a81fc8aff73b1982466c614e359364c49f01ca7
Reviewed-on: https://gerrit.causes.com/24885
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>