Move commit-stats -> twig-stats
This is in further preparation for gemmification; our binaries will get
installed in the PATH along with whatever else might be in there, so we
need something in the name to differentiate them.

For now that means `twig-app` and `twig-stats`, although I'd like to
turn these into subcommands of a single `twig` executable as a
subsequent step.

This commit renames `commit-stats` to `twig-stats` and updates the
references in the doc accordingly. I've already dropped the `bin/`
prefixes from the names, as they won't be needed any more once this is a
gem.

Change-Id: Iad48ea5babd80702f7c52322689ffb21851f5767
Reviewed-on: https://gerrit.causes.com/24886
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>