Add favicon files
For now, we're just using the Causes favicon. We've added a PNG with the
icon, and also an empty favicon.ico file, because browsers tend to
request that as a matter of course anyway, and having something to serve
them up keeps 404 noise out of the logs.

There is some un-DRY-ness in the templates, which we will fix shortly by
adding a layout.

Change-Id: Ib9e5837fff4f3d395058d45d6e49060ef76436ce
Reviewed-on: https://gerrit.causes.com/24889
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>