Add files from Bundler gem template
We're going to be releasing this project as a gem, so in preparation,
here we add the files generated by `bundle gem twigg`, with some tweaks
(replaced double quotes with singles, added some docs, removed the
dependency on Git, switching to autoloading of the version number etc).

It still is not suitable to ship as a gem (for example, we need to
rework the handling of the config file) but it is a necessary
precondition.

Change-Id: I41418e859b8dd1d4c19a29dcb2165c9983c31b43
Reviewed-on: https://gerrit.causes.com/24893
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>