Show large numbers with delimiters
For now this is done via a new Twigg::Util module. The module extends
itself, which means that you can either mix-in the methods and use them
directly, or call `Twigg::Util.<method>`.

In one place that I'm using it, I made a wrapper method as a
convenience, to avoid importing the module and expanding the public
interface of the class. I might revise that decision later on, as I am
no sure whether the rigor is worth sacrificing the convenience for.

Change-Id: I8381e88b84040a5a4e770e0a59132817b3bb8d20
Reviewed-on: https://gerrit.causes.com/24914
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>