Adapt config handling to work in a gem context
If people install Twigg as a gem, we can't expect them to place their
config file at a fixed location relative to the gem.

This commit adds a couple of new classes, Twigg::Config and
Twigg::Settings, that encapsulate the concern of finding and accessing
config information.

With this, you can either place the config file at a standard location
(eg. `~/.twiggrc`) or you can pass in a custom location via the
`TWIGGRC` environment variable.

In a later commit, we'll probably add a `-c` or `--config` switch that
allows you to supply a path to a config file from the command-line as
well.

For now, the implementation is relatively crude, with easy of use for
API clients foremost. If we later find ourselves needing to support
things like nested configuration or other fanciness, we can cross that
bridge when we come to it.

Change-Id: I30bdc428c1c3a2e09e4b5aea18b13f59f58708ad
Reviewed-on: https://gerrit.causes.com/24915
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>