Move some common option handling up into Command class
We can support -v/--verbose and -h/--help for all subcommands by putting
the code for it into the Command class, which is the superclass of the
various subcommand classes.

Note that "support" for -v/--verbose means that we set a `@verbose`
instance variable; the onus is on the subcommands to actually modify
their behavior acccordingly.

Change-Id: I511cbec701b4663f93615198930feb6554ac51e3
Reviewed-on: https://gerrit.causes.com/25238
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>