Add first cut of "gerrit" subcommand
This is just a crude first cut. It basically works; you can clone and
update repos with it.

The `--verbose` switch is ignored because I'm setting the value in the
superclass, which instantiates an instance of the subclass, and that
instance variable is obviously not shared. I'll fix that in a later
commit.

I can see this subcommand evolving sub-subcommands of its own in the
future. For example:

  twigg gerrit clone # clone uncloned repos
  twigg gerrit update # update all repos
  twigg gerrit stats # show stats (changesets, comments, reviews etc)

Change-Id: I5970ab2f860ef6dbedded479e1c6e95679cd8838
Reviewed-on: https://gerrit.causes.com/25245
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>