Prepare Command class for proper inheritance
As noted in prior commits, we're not making proper use of inheritance in
the Command subclasses. Basically, there is common code in the
superclass `#initialize` method that sets instance variables, but we
can't access them from the subclasses because they don't call `super`.

This commit paves the way for us to be able to do that by removing all
the stuff from the `Command#initialize` method that depends on instance
methods. We'll then be able to move the remaining stuff into a class
method, which itself will initialize the subclasses, and they'll be able
to call `super` and get access to the instance variables and share the
common initialization without repetition.

To make this happen:

  - move `#usage` down into the Help class, as a faux-sub-subcommand
  - make `#die` a public method, which means we can call it on the
    returned `Help` instance (the alternative of having Help methods
    automatically die once they've done their thing was eschewed in
    favor of having something easy to test)

Change-Id: Ief5188637e7ffd73155475a1533d745d4f71c2a7
Reviewed-on: https://gerrit.causes.com/25251
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>