Complete the inheritance overhaul for the Command class
This commit completes the work started in the previous commit; it moves
logic out of the Command#initialize method and up into a new Command.run
method. Several support methods were also promoted to class methods.
Added delegation so that the methods that still need to be used as
instance methods delegate to the class methods.

Note: there should be no behavioral change in this commit; it's just a
refactoring. In the next commit I will rework the subclasses to leverage
the functionality of their superclass.

Change-Id: I7288f0884563b6f39aca4916fd1945b4f53357f0
Reviewed-on: https://gerrit.causes.com/25252
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>