Remove a couple of `#die` calls
One of these we turn into a graceful exit (passing `-h` or `--help`
should just show the help without employing a non-zero exit code).

Another, the `Help#usage` method, we get rid of the `die` call, as the
caller does this for us via a call to `#run!` (which calls `#run` then
`#die`). This is good because it makes the `#usage` method more easily
testable.

Change-Id: I3a0e286e495a4049cd1a501405314337f3cd3d46
Reviewed-on: https://gerrit.causes.com/25254
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>