Remove unnecessary temporary variable in Command.run
I had originally assigned a value to the `debug` local variable in
`Command.run` to protect it from getting destroyed when calling the
subcommands (ie. which all call `super`, which ends up deleting it from
the args array).

It turns out we don't have to worry about this:

  args.include?('-d') # present
  some_method_which_deletes_an_item_in_args(*args)
  args.include?('-d') # still present

The value survives because the splat effectively creates a new "array",
and when `#some_method_which_deletes_an_item_in_args` mutates the
"array" (really a method parameter list) the original is untouched.

Change-Id: Iee05cef809af743e931713f3d8122dd840a3b24d
Reviewed-on: https://gerrit.causes.com/25256
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>