Fix a couple of subtle bugs in Settings magic
As clever is the new dumb, it should come as no surprise that my
"magic" in the Settings class had some bugs in it.

Firstly, we were assigning uselessly to the `_namespace` local variable.
This variable was only ever `true`, and passing it down into the
`override_methods!` call was, albeit harmlessly, constructing instance
variable names like `@true__user` instead of the desired `@gerrit_user`.

Secondly, and more nefariously, we observed that accessing values like
`Config.gerrit.user` worked perfectly on the first access, but returned
`nil` for subsequent accesses.

The culprit here was the line:

  return value if value = ...

I am not entirely sure why, but this invalid Ruby, which would normally
explode, was buzzing along quite happily in the context of this method;
perhaps there is some closure magic going on, or, more likely, the fact
that this method is being defined dynamically within a block is
suppressing the normal warning and error behavior.

A simple test shows that Ruby doesn't consider this pattern legit:

  def foo
    return bar if bar = true
  end

This method doesn't return `true`, but rather first warns:

  warning: found = in conditional, should be ==

at parse time, and then raises at runtime:

  NameError: undefined local variable or method `bar' for ...

Change-Id: I69638c9490b0c6b7ae3a36ecbb98f9db47cd2677
Reviewed-on: https://gerrit.causes.com/25257
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>