Remove _namespace hack
We were using a `_namespace: true` key value pair as a signal that a
hash was a namespace and not a mere options hash. We can do this more
elegantly by just seeing if a hash is multi-level.

This means I can delete the comment about it being impossible to ever
have a setting called "_namespace" (although it's unlikely we'd ever
want that).

More importantly, it will come in handy when I remove the limitation
that namespaces cannot be nested. We don't have a specific need for
multi-level nesting yet, but the limitation is arbitrary and the code
will arguably be cleaner without it.

Change-Id: I521ef4372b382dded5a27f60e893d165c557cbcf
Reviewed-on: https://gerrit.causes.com/25261
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>