Add support for nested Settings namespaces
Pleasingly, this change doesn't even require us to modify the code that
runs at init time. All we have to do is replace our uni-level instance
variable with a stack, and prepare the nested hashes accordingly. This
will even support re-opening namespaces and adding additional settings
to them, if we ever want to.

There may be an element of YAGNI in all this, but I think the resulting
code is cleaner and more congruent.

Change-Id: I8d081cb5a694d82f32e4fb13972a9e12c56d1d3e
Reviewed-on: https://gerrit.causes.com/25262
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>