Replace File.dirname(__FILE__) calls with __dir__
As we're on Ruby 2.0, the latter is a convenient shortcut which we
should use. Note I do not use it in the gemspec, because it is possible
that that may be evaluated by an older version of Ruby (for example, an
older Ruby might read this file and blow up on hitting `__dir__` and
before getting to the `required_ruby_version` setting).

Change-Id: I4dfa2b4aad53b1047c322fd40fac5a59babcf59a
Reviewed-on: https://gerrit.causes.com/25265
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>