Move bulk of Twigg::Settings into Twigg::Settings::DSL
The whole point of this DSL is to produce an easily readable
specification of the different options, defaults and validations. Having
the implementation of the DSL in the same file undermines that goal,
because it is so verbose and overwhelming compared to the list of
settings themselves.

So, extract the DSL code out into a separate module. This is clearly
somewhat sucky, as it somewhat arbitrarily disperses the concerns of the
class between the class and the module. Nevetheless, on the balance I
think it's worth it in terms of the readablity benefit for the Settings
class. (I expect we'll be adding to the Settings file over time, whereas
the DSL code is pretty much done at this point and much less likely to
be touched.)

Change-Id: I16a90d2b074024252f235d0fb0ea42ea487392fc
Reviewed-on: https://gerrit.causes.com/25275
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>