Teach Settings DSL about required settings
I'm soon going to be adding GitHub support and some of those settings
are going to be required (there aren't sensible defaults that we could
provide).

So, this commit adds a `required: true` option to the DSL and uses it in
the one place where we already have "required with no sensible default"
dynamics.

Change-Id: I99b47d4162bfe20553442240f686d66506a09fbd
Reviewed-on: https://gerrit.causes.com/25276
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>