Add GitHub subcommand, `twigg github`
This commit adds support for cloning/updating repos from GitHub via a
new `twigg github` subcommand (the sub-subcommands are `twigg github
clone` and `twigg github update`). It is very similar to the `twigg
gerrit` subcommand (and in fact in a subsequent commit I will be moving
a large chunk of identical code up into a common superclass that both of
these subcommands can use).

Hitting the GitHub API is somewhat painful from Ruby, due to the way it
handles SSL certificates. I could either disable verification (a bad
idea), or jump through a hoop to get it working; I ended up jumping
through the hoop by bundling the CA certs and the cert itself in a
little chain and including it with Twigg.

The PEM file was created by going to https://api.github.com/ in Chrome,
clicking on the padlock in the location bar, and in the certificate
details view dragging and dropped each certificate to the desktop; from
there I was able to use the command line:

   openssl x509 -outform pem -out a.pem -inform der -in DigiCert\ High\ Assurance\ EV\ Root\ CA.cer
   openssl x509 -outform pem -out b.pem -inform der -in DigiCert\ High\ Assurance\ CA-3.cer
   openssl x509 -outform pem -out c.pem -inform der -in \*.github.com.cer
   cat {a,b,c}.pem > github.pem

I got the idea for this from:

  https://gist.github.com/mislav/5025114

But there are many different hits detailing various strategies for
dealing with the fact that Ruby doesn't behave much like a browser, or
like OpenSSL on the command line; a couple of examples:

  http://stackoverflow.com/a/9238221
  http://jackassofalltrades.org/2012/02/rubys-xmlrpcclient-and-ssl/

Working with the Net::HTTP library is a bit of a pain at the best of
times, so I might look for something simpler in the future. Another
thing I'd like to improve is the handling of pagination of API requests;
the parsing of the `Link` header feels a little fragile.

Change-Id: I5496806b84e0bb83c297c05c99c2e6f010ee408d
Reviewed-on: https://gerrit.causes.com/25278
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>