Move common GitHub/Gerrit functionality into superclass
This commit creates a `GitHost` class which is an abstract superclass to
the `Gerrit` and `GitHub` classes. It DRYs up a bunch of repeated
functionality by moving it up the inheritance hierarchy.

This is just a straight move of all the methods. The only one that
needed tweaking was the `#initialize` method, where I replaced a couple
of hard-coded strings with introspected ones (eg. "gerrit" and "github"
got replaced with `self.class.to_s.split('::').last.downcase`).

Change-Id: I9fb5722a18ed394f69a50580aadeadfee6be0dd8
Reviewed-on: https://gerrit.causes.com/25280
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>