Warn if TWIGGRC has loose filesystem permissions
As the config file may contain things like GitHub access tokens (which
are effectively almost as useful as passwords), we should warn if we
note that it has weak permissions on it.

I made some of the utility methods on the Command class public so that I
could use them in this commit. It is somewhat inelegant having to talk
to the Command class for these things, so I will likely move them into a
module in a subsequent commit.

Change-Id: Ie9f5181342d605e8466b4cb24af8194d44d51b1f
Reviewed-on: https://gerrit.causes.com/25285
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>