Add specs for Settings class
These specs are partly a verification of the Settings::DSL
functionality, and partly a verification that I've made the right calls
to the DSL in the Settings class itself. For this reason, I've tried to
walk a middle ground of testing general features (stuff around deep
OpenStructification) and testing specific configuration (for example,
making use of the `default_days` and `repositories_directory` settings).

I may follow up later with some more generic tests that are isolated
from the set-up created by the DSL in the Settings class; for example,
it would be nice to have some tests that show that namespaces can be
nested arbitrarily deeply.

Change-Id: I5bb3e8e1eeb3bdb5472c3cfdc8784bc2e934cbc0
Reviewed-on: https://gerrit.causes.com/25287
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>