Finishing move of lib specs into specs/lib/
This was originally attempted in commit cbd57c7, but Gerrit
unfortunately only cherry-picked half of the changeset; you can see the
original changeset as it was first pushed here:

  https://gerrit.causes.com/25288

Despite what the docs[1] suggest, you can't really submit a bunch of
changes at once via the command line; Gerrit will effectively merge them
in arbitrary order. In this case, rather than producing a merge
conflict, it just decided to merge only half of the change.

For future reference, the way to do this is us a `for` loop in the shell
to issue one "submit" command at a time to the Gerrit server.

In the meantime, this commit supplies the missing half of the broken
commit.

[1]: https://gerrit.causes.com/Documentation/cmd-review.html

Change-Id: I63aa81db4871f376a4eaafcc267ff221a10c7005
Reviewed-on: https://gerrit.causes.com/25290
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>