Add specs for Twigg::Settings::DSL, fixing a latent bug
As mentioned in the previous commit, we're already tested part of this
functionality in the Twigg::Settings specs. Those specs are partly tied
to the settings hierarchy as it is specifically manifested in the
Twigg::Settings class, by virtue of the calls it makes to the DSL.

The specs added here are more generic in nature. I create anonymous
classes (to avoid possible future naming clashes) and make up some phony
settings for the purposes of exercising three capabilities of the DSL
that are not exercised in the other specs:

  - nesting namespaces
  - re-opening pre-existing namespaces and adding more settings
  - memoization

We're not using either of these features yet, but I wanted to have them
reflected in the spec suite because the system was designed with them in
mind.

I was delighted to discover a latent bug in the course of writing these
tests; we weren't running into it in our use, but the spec revealed that
it was possible for `@namespaces` to be `nil` when the first use of the
DSL was a `settings` call instead of a `namespace` one. Fix that by
mediating access to the instance variable through a new reader method,
`#namespaces`.

These are the first specs that need mocking. I went with RR despite the
fact that we have mostly used Mocha in other projects at Causes and
people are familiar with it; RR offers a range of powerful doubles
(including spies and proxies) that Mocha doesn't support.

There is a deprecation warning with the version of RSpec that I'm using:

  DEPRECATION: RSpec::Core::Configuration#backtrace_clean_patterns is
  deprecated. Use
  RSpec::Core::Configuration#backtrace_exclusion_patterns instead.

but this is already fixed in this pull request, which should be merged
soon:

  https://github.com/rr/rr/pull/37

Change-Id: Id2925766e89e46273eef804c0088c353f2779af3
Reviewed-on: https://gerrit.causes.com/25289
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>