Escape shell words before showing them to the user
If anybody ever has something nefarious like a space in the path to
their TWIGGRC file or their Gem install path, this commit ensures that
we don't end up displaying a command for them to run that won't actually
work if they paste it in the shell.

For example, if TWIGGRC is `/foo/bar baz/twiggrc`, we would here print
`/foo/bar\ baz/twiggrc` in our permissions checking code, which can
safely be copied and pasted into the shell.

Change-Id: I3dfc25a75df92b25deff982635445916132a2153
Reviewed-on: https://gerrit.causes.com/25292
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>