Don't die on seeing commits with invalid timezones
rails.git has some commits like 4cf94979c9f4d6683c with invalid
timezones on the committer and author lines; eg.

  committer Vijay Dev <vijaydev.cse@gmail.com> 1312735823 +051800

This is supposed to be a 4-digit number preceded by + or -, but in this
case is a 6-digit number, and a non-sensical one at that.

As we effectively ignore the timezone info anyway (we're initializing
the date based on the number of seconds elapsed since the UNIX epoch,
which is always expressed in GMT/UTC), this commit adjusts the commit
parser to not choke on the invalid timezone.

Change-Id: I1556b21d2956cdaf63905cdc81ebed109a7dcac6
Reviewed-on: https://gerrit.causes.com/25294
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>