Handle commits with invalid author info
We were barfing on bad commits like commit 4e873ffcdab0c445 in
rails.git, which has an author line looking like:

  author  <JRadosz@gmail.com> 1302595195 -0700

ie. no author name, just an empty string.

This commit teaches Twigg to grab both the name and the email, so that
we can fallback to the email if the name is not present. With this
change, we can parse the last 1,000 days of rails.git history without
barfing; I'll follow up with an even longer scale test after this.

Change-Id: Ibf0f1eb68dfa679c0b10164942f59001d111c4ca
Reviewed-on: https://gerrit.causes.com/25295
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>