Add per-author aggregate stats in `stats` verbose mode
This method is long and ugly. I'll look at cleaning it up as I enhance
the handling of width calculations (ie. to make them dynamic, rather
than assuming a maximum width).

Will need to make the same change for the web app if we want to keep it
at parity with the CLI version.

Change-Id: I84ff3414651115e57e70a341ede12cdd886f4483
Reviewed-on: https://gerrit.causes.com/25299
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>