Teach web app to link back to GitHub for project names
In order to facilitate this, I added a `#link` method to `Twigg::Repo`,
that behaves like its counterpart on `Twigg::Commit` does (ie. if
`github.organization` is set, it returns a link; otherwise it returns
`nil`).

Likewise, I adapted to `#count_by_repo` to use real `Twigg::Repo`
objects as keys, so that I could call the `#link` method as needed to
construct the URL.

Change-Id: I622caf5dd4929ac70922a1510aafcb1a8c353931
Reviewed-on: https://gerrit.causes.com/25301
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>