Move twigg files into subdirectory
As Twigg grows, and we add more features, we bring in more dependencies.
For example, we currently depend on Sinatra and Haml. Soon, we'll
depend on Sequel and likely mysql2 as we add the `gerrit stats` feature.

In order to prevent the gem from becoming bloated, we're going to split
it up into multiple gems. The main gem will be "twigg" and will depend
on nothing. Then we'll add "twigg-app", which will depend on "twigg" and
"sinatra". Later will come "twigg-gerrit", which will depend on "twigg"
and "mysql2" (likely) and "sequel".

We'll teach `twigg` to degrade gracefully when the various component
gems aren't available, but if you want everything, you'll be able to:

  gem install twigg twigg-app twigg-gerrit # etc

This commit is the first step in that process; we move all of the files
into a "twigg" subdirectory. These will form the basis of the "twigg"
gem. The next step will have us extracting the app files out into
"twigg-app".

Change-Id: Iac274836fd09ea376355adea87a8af14649b51b3
Reviewed-on: https://gerrit.causes.com/25816
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>