Extract twigg-app gem
Move most the app files out of "twigg" into "twigg-app", and make the
later depend on the former in the gemspec (for easy installation) and
the former depend on the later in the Gemfile (for development).

In order for `bundle check` to work in both subdirectories, we manually
add the relative path to the other subdirectory in the "twigg-app"
Gemfile as well.

Other than that, this is mostly a straight move. Added `Twigg::App.root`
so that the code in the "twigg-app" gem could also have a convenient way
of finding its root directory, and added `Twigg::App::VERSION`.

Change-Id: I3d6013a9298bc2ea1c38e3b882d6ef9f31252854
Reviewed-on: https://gerrit.causes.com/25818
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>