Extract twigg-gerrit gem
Much as we did with the twigg-app gem, extract the Gerrit functionality
into a gem, twigg-gerrit. We're doing this in order to keep the number
of dependencies of the main twigg gem small; if you want the Gerrit
functionality, you install the twigg-gerrit gem.

For now, this won't make much of a difference, but it will matter more
shortly, because we're going to be adding at least the sequel gem, and
likely the mysql2 gem as well, for the purposes of the "gerrit stats"
subcommand.

Most of this commit is just a straight move. Added a
`Command#with_dependency` method to encapsulate the pattern of working
with optional dependencies. I also needed to add a couple of `require
'forwardable'` calls, because the load ordering with this new set-up
meant that it wasn't pre-loaded by the time the code that was using it
needed it.

Change-Id: I672b478a66e9d8e7a81488c4f2305870d9346d7a
Reviewed-on: https://gerrit.causes.com/25822
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>