Add sequel as a dependency to twigg-gerrit
We make sequel a hard dependency. mysql2 is only a "soft" dependency
because it is possible that people may want to use different adapters
with it (for example, mysql or postgres).

Note that the "twigg/Gemfile" still changes here, because in the
development environment (only) "twigg" depends on the "twigg-gerrit"
(and "twigg-app") repos, and so drags in all their dependencies as well.

This will be used in a forthcoming commit to mediate access to the
Gerrit database.

Change-Id: Ice212b00c694b79b958349979292487c422cb50a
Reviewed-on: https://gerrit.causes.com/25824
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>