Add -c/--config switches
As noted in the comments, there is an unfortunate smell here that we
have little choice but to work around.

In "production", most (or all?) interaction will occur via the `twigg`
executable, and that means the Twigg::Command.run method, which can
process the options.

In development, crazy things happen, even when running the `twigg`
executable. First of all, we use Bundler in development, which looks
at the gemspec but also the Gemfile, and so ends up evaluating the
gemspec of dependent gems like twigg-app. This in turn forces the App
class to load (because we need to access `Twigg::App::VERSION`) even
before we get to our Twigg::Command.run call, and the class-level DSL
there ends up invoking Twigg::Config.config before we've had a chance to
process the arguments.

The only way to defend against this insanity is to teach the
Twigg::Config class to process `-c`/`--config` options. We put the logic
for processing options in the Twigg::Console class, where it is not so
out of place, and we make sure that Twigg::Command.run at least calls
Twigg::Config.config to ensure the options get processed.

On the bright side, this implementation is robust enough to work even
with things like `bundle exec shotgun config.ru` in development, which
bypass the `twigg` executable entirely.

It's ugly, but a necessary evil in order to support this feature.

Change-Id: I79ef80181bcdf121bc039a26f5302fea77eb1f9d
Reviewed-on: https://gerrit.causes.com/25827
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>