Go back to reading entire input in one fell swoop
As mentioned in the previous commit, switching to a "lazy" enumerator
based approach actually made things slower:

  scanner (3 sample runs of 3,000 days of history across 113 repos):
  63 secs
  65 secs
  62 secs

  enumerator
  113 secs
  119 secs

This commit goes back to reading the entire file at once. We're still
using an enumerator, but we're no longer passing blocks around:

  enumerator, but reading entire file at once
  63 secs
  60 secs

This is marginally faster than what we had before, but the difference is
small enough to be classified as noise.

Change-Id: I38ef8a1fc2af7c306174b6a25413e92b7db10a78
Reviewed-on: https://gerrit.causes.com/25830
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>