Handle -c/--config option application better
Instead of waiting until the `.run` method is called, we can do better,
and ensure that the -c/--config option is applied before any other, by
doing it in the top of the `#initialize` method. Note that this also
cleans up some sloppiness, where we could potentially consume the option
from `ARGV`, but not from the `args` array, depending on the order of
evaluation.

It still isn't the prettiest code ever, but it is at least in a better
place now.

Change-Id: I51918b3d801e7c17eba9a9f5863fa5afbdbef659
Reviewed-on: https://gerrit.causes.com/25831
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>