Add "teams" setting
This one has a default value of `{}`, which means we need to rework our
Namespace implementation (which formerly relied on the ability to check
for Hash-ness in order to identify namespaces).

Here we make a Hash subclass for the sole purpose of indicate
namespace-ness.

A future commit will make use of this setting to show team-centric
stats.

Change-Id: I6d8c3e2eecb7859b8ffb75a838107c9ca3af84ef
Reviewed-on: https://gerrit.causes.com/25834
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>