Add crude first cut of /teams page
As an initial starting point this is just a copy of the /authors page,
but grouped by team, with substantial portions of the template copied.

The end goal here is for this view to be truly team-centric (ie. the
individual authors shouldn't be mentioned on the index page; that's what
the /authors page is for), but for now this is just a means of getting
something on the page for subsequent iterations.

Change-Id: I805e676a5c74de66a9309e99640ab124b0a3e26b
Reviewed-on: https://gerrit.causes.com/25839
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>