Show team member names in tool tips
As a step towards making the /teams page more team-centric, add
tool-tips showing the team member names. This in turn will mean that we
can remove the rows that show the team members, which will allow us to
emphasize teams over members.

Required some rejigging of the JS in order to make the tooltip code from
Bootstrap available on every page.

Change-Id: I34724e06c2f031aba50dac070b174ab45c6c5781
Reviewed-on: https://gerrit.causes.com/25841
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>